refactoring + added hydra config
