+++ src/train.py
@@ -1,30 +1,34 @@
-import logging
-import sys
-from logging.handlers import RotatingFileHandler
+from omegaconf import DictConfig, OmegaConf
+import hydra
 
-import torch.nn as nn
-from constants import (
-    NOUN_CLASSES,
-    VERB_CLASSES,
+from systems.recognition_module import (
+    EpicActionRecognitionDataModule,
+    EpicActionRecognitionModule,
 )
-from systems.trainer import Trainer
 from utils import get_loggers
 
-# LOGGING
-stream_handler = logging.StreamHandler(stream=sys.stdout)
-file_handler = RotatingFileHandler(
-    filename="data/pilot-01/logs/train.log", maxBytes=50000, backupCount=5
-)
-logger = get_loggers(
-    name=__name__,
-    handlers=[(stream_handler, logging.INFO), (file_handler, logging.ERROR)],
-)
+LOG = get_loggers(name=__name__, filename="data/pilot-01/logs/train.log")
+
+
+@hydra.main(config_path="../configs", config_name="pilot_config")
+def main(cfg: DictConfig):
+    LOG.info("Config:\n" + OmegaConf.to_yaml(cfg))
+    data_module = EpicActionRecognitionDataModule(cfg)
+    LOG.debug("EpicActionRecognitionDataModule initialized")
+    system = EpicActionRecognitionModule(cfg, data_module)
+    LOG.debug("EpicActionRecognitionSystem initialized")
+    if not cfg.get("log_graph", True):
+        # MTRN can't be traced due to the model stochasticity so causes a JIT tracer
+        # error, we allow you to prevent the tracer from running to log the graph when
+        # the summary writer is created
+        try:
+            delattr(system, "example_input_array")
+        except AttributeError:
+            pass
+    LOG.info("Starting training....")
+    system.training_loop(1, cfg.model.save_path)
+    LOG.info("Training completed!")
+
 
 if __name__ == "__main__":
-    trainer = Trainer(
-        VERB_CLASSES,
-        NOUN_CLASSES,
-        nn.CrossEntropyLoss(reduction="mean"),
-        save_path="data/pilot-01",
-    )
-    trainer.training_loop(num_epochs=1)
+    main()
